requirements.txt
antlr4-python3-runtime==4.7.2
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
asttokens==1.1.13
autopep8==1.4.3
backcall==0.1.0
baron==0.9
bleach==3.0.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
comtypes==1.1.7
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.2.1
docopt==0.6.2
entrypoints==0.2.3
Flask==1.0.2
future==0.17.1
gevent==1.4.0
gitdb2==2.0.5
GitPython==3.0.2
GPIOSimulator==0.1
greenlet==0.4.15
grip==4.5.2
idna==2.7
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
langdetect==1.0.7
lazy-object-proxy==1.4.2
locustio==0.11.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.4
modernize==0.7
msgpack==0.6.1
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
notebook==5.7.0
numexpr==2.6.9
numpy==1.15.2
opencv-python==3.4.3.18
orphanblack==0.2.0
pandas==0.23.4
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.3.1
path-and-address==2.0.1
pexpect==4.6.0
pickleshare==0.7.5
pika==1.1.0
Pillow==5.3.0
pipenv==2018.11.26
plotly==3.7.1
ply==3.11
pprint==0.1
prometheus-client==0.4.1
prompt-toolkit==2.0.5
psutil==5.6.1
ptyprocess==0.6.0
pyaml==19.4.1
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.2.0
pylint==2.3.1
pyOpenSSL==18.0.0
pyparsing==2.2.2
PyRSMQ==0.4.1
python-compiler==1.1
python-dateutil==2.7.3
python-dotenv==0.10.3
pytz==2018.7
PyYAML==5.1.2
pyzmq==17.1.2
redbaron==0.9.2
redis==3.3.8
requests==2.19.1
retrying==1.3.3
rply==0.7.7
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
SimpleWebSocketServer==0.1.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smmap2==2.0.5
sqlparse==0.3.0
stdlib-list==0.5.0
stop-words==2018.7.23
tabulate==0.8.3
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typed-ast==1.4.0
Unidecode==1.0.22
urllib3==1.23
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.2
xlrd==1.2.0
